Main player in main combat now has attack speed, and attack speed bar is shown

+++ DoubleM/Assets/Scripts/Healthbar.cs
@@ -9,7 +9,7 @@ public class Healthbar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
@@ -17,7 +17,7 @@ public class Healthbar : MonoBehaviour
         fill.color = gradient.Evaluate(1);
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
